+++ src/components/DocPages/Container.jsx
@@ -1,5 +1,4 @@
 import styled from "styled-components"
-import * as styleVars from "../variable"
 
 export const Container = styled.section`
   display: flex;