+++ docs/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1](https://github.com/hackmcgill/mchacks7/tree/1.3.1) - 2019-12-21
+
+### Added
+
+- Add google analytics
+- Add MLH 2020 badge
+
+### Fixed
+
+- Fix compiler warnings
+
+### Security
+
+- Update to Gatsby v2.18.16 to resolve dependency warnings
+
 ## [1.3.0](https://github.com/hackmcgill/mchacks7/tree/1.3.0) - 2019-12-17
 
 ### Added